+++ integration-tests/api/__tests__/store/orders.js
@@ -168,6 +168,7 @@ describe("/store/carts", () => {
         // fields
         "status",
         "email",
+
         // relations
         "shipping_address",
         "fulfillments",
@@ -177,7 +178,8 @@ describe("/store/carts", () => {
         "customer",
         "payments",
         "region",
-        // default
+
+        // totals
         "shipping_total",
         "discount_total",
         "tax_total",
@@ -199,6 +201,7 @@ describe("/store/carts", () => {
       expect(Object.keys(response.data.order)).toEqual([
         // fields
         "status",
+
         // default relations
         "shipping_address",
         "fulfillments",
@@ -208,7 +211,8 @@ describe("/store/carts", () => {
         "customer",
         "payments",
         "region",
-        // default
+
+        // totals
         "shipping_total",
         "discount_total",
         "tax_total",
@@ -232,9 +236,11 @@ describe("/store/carts", () => {
       expect(Object.keys(response.data.order)).toEqual([
         // fields
         "status",
+
         // selected relations
         "billing_address",
-        // default
+
+        // totals
         "shipping_total",
         "discount_total",
         "tax_total",