+++ packages/admin-ui/ui/src/components/templates/settings-overview.tsx
@@ -1,5 +1,6 @@
 import React from "react"
 import PageDescription from "../atoms/page-description"
+import Spacer from "../atoms/spacer"
 
 const SettingsOverview: React.FC<React.PropsWithChildren> = ({ children }) => {
   return (
@@ -11,6 +12,7 @@ const SettingsOverview: React.FC<React.PropsWithChildren> = ({ children }) => {
       <div className="medium:grid-cols-2 gap-x-base gap-y-xsmall grid auto-cols-fr grid-cols-1">
         {children}
       </div>
+      <Spacer />
     </div>
   )
 }