+++ packages/medusa/src/api/routes/admin/orders/refund-payment.ts
@@ -9,6 +9,7 @@ import {
 import { OrderService } from "../../../../services"
 import { EntityManager } from "typeorm"
 import { FindParams } from "../../../../types/common"
+import { cleanResponseData } from "../../../../utils/clean-response-data"
 
 /**
  * @oas [post] /admin/orders/{id}/refund
@@ -97,7 +98,7 @@ export default async (req, res) => {
     includes: req.includes,
   })
 
-  res.status(200).json({ order })
+  res.status(200).json({ order: cleanResponseData(order, []) })
 }
 
 /**