+++ packages/medusa/src/models/line-item-adjustment.ts
@@ -8,10 +8,9 @@ import {
   PrimaryColumn,
 } from "typeorm"
 
-import { DbAwareColumn } from "../utils/db-aware-column"
+import { DbAwareColumn, generateEntityId } from "../utils"
 import { Discount } from "./discount"
 import { LineItem } from "./line-item"
-import { generateEntityId } from "../utils/generate-entity-id"
 
 @Entity()
 @Index(["discount_id", "item_id"], {
@@ -41,7 +40,7 @@ export class LineItemAdjustment {
   @Column({ nullable: true })
   discount_id: string
 
-  @Column({ type: "int" })
+  @Column({ type: "numeric", transformer: { to: (value) => value, from: (value) => parseFloat(value) } })
   amount: number
 
   @DbAwareColumn({ type: "jsonb", nullable: true })
@@ -93,7 +92,7 @@ export class LineItemAdjustment {
  *     $ref: "#/components/schemas/Discount"
  *   amount:
  *     description: The adjustment amount
- *     type: integer
+ *     type: number
  *     example: 1000
  *   metadata:
  *     description: An optional key-value map with additional details