+++ packages/medusa/src/services/totals.ts
@@ -58,6 +58,8 @@ type LineItemTotals = {
   original_tax_total: number
   tax_lines: LineItemTaxLine[]
   discount_total: number
+
+  raw_discount_total: number
 }
 
 type LineItemTotalsOptions = {
@@ -460,7 +462,7 @@ class TotalsService extends TransactionBaseService {
             /**
              * Used for the refund computation
              */
-            unit_amount: Math.round(adjustmentAmount / ld.item.quantity),
+            unit_amount: adjustmentAmount / ld.item.quantity,
           }
         } else {
           allocationMap[ld.item.id] = {
@@ -795,7 +797,8 @@ class TotalsService extends TransactionBaseService {
       subtotal = 0 // in that case we need to know the tax rate to compute it later
     }
 
-    const discount_total = lineItemAllocation.discount?.amount ?? 0
+    const raw_discount_total = lineItemAllocation.discount?.amount ?? 0
+    const discount_total = Math.round(raw_discount_total)
 
     const lineItemTotals: LineItemTotals = {
       unit_price: lineItem.unit_price,
@@ -807,6 +810,8 @@ class TotalsService extends TransactionBaseService {
       original_tax_total: 0,
       tax_total: 0,
       tax_lines: lineItem.tax_lines || [],
+
+      raw_discount_total,
     }
 
     // Tax Information
@@ -1002,7 +1007,9 @@ class TotalsService extends TransactionBaseService {
       excludeNonDiscounts: true,
     })
 
-    const discountTotal = this.getLineItemAdjustmentsTotal(cartOrOrder)
+    const discountTotal = Math.round(
+      this.getLineItemAdjustmentsTotal(cartOrOrder)
+    )
 
     if (subtotal < 0) {
       return this.rounded(Math.max(subtotal, discountTotal))